refactor: Resolve symfony/console deprecation in 7.4, fix phpstan
symfony/console 7.4.0 deprecates the `Application::add` method in favour
of a new `Application::addCommand` method. `Application::add` is then
removed in 8.0.

We had a fix for this in the 4.0 branch, but it started to cause phpstan
errors once symfony 8.0 was released & installed for the phpstan job.
Additionally, the errors varied depending on the symfony version
installed.

This fix bumps us off the deprecated method for Behat 3.x and symfony >=
7.4.0. It also updates the phpstan ignore to cover all supported symfony
versions.