ci: Update baseline for expected BC breaks in 4.0 so far
Now that we are running BC checks in CI, we will need to update the
baseline in subsequent 4.x PRs (if they contain breaking changes).

The baseline will give us a good reference to check that we have
documented all BC breaks in the changelog / upgrade guide when we come
to prepare the initial 4.0 series release(es).