[3.x] refactor: Improve type-safety of regex string replacement operations (#1778)
While starting work on the next iteration of adding strict return types,
Rector picked up a number of methods that had phpdoc of `@return string`
but could theoretically also have been returning `null`.

This is because `preg_replace` and `preg_replace_callback` can return
null in some cases (e.g. if the pattern is valid but it hits the runtime
backtrack_limit). We were not always handling this - sometimes casting
back to empty string, sometimes just assuming the result would be a
string.

This can also happen if there is an issue with the regex itself (e.g.
the pattern is invalid) and the runtime PHP error is not captured by an
error handler that converts it to an exception - generally in Behat we
do, but the static analyser cannot detect this.

This commit introduces thin wrapper functions to ensure that the
preg_replace / preg_replace_callback was successful. This will reduce
the need for assertions elsewhere in the code as we move towards
stricter typing.

Note that the native PHP functions can also return an `array` if passed
an `array $subject` however we never use this feature so I have limited
the utility functions to `string $subject` and ` string` return type. If
we needed to support arrays in future I'd be inclined to do that as
separate methods (e.g. `replaceArray`) so that we only have to handle
array return types in places we expect them.